Make various UX improvements to the CLI (#187)
* Enable `--parallel` by default & consolidate its definition

* Add `plt add-repo` subcmd, and alias `require-repo` to `add-repo`

* Make `plt clone` cache required repos afterwards

* Cache all required repos after `[dev] plt add-repo` and `plt pull`

* Add `stage set-next-result` subcommand

+++ cmd/forklift/plt/pallets.go
@@ -13,12 +13,12 @@ import (
 )
 
 func processFullBaseArgs(
-	c *cli.Context, ensureCache bool,
+	wpath string, ensureCache bool,
 ) (pallet *forklift.FSPallet, cache forklift.PathedRepoCache, err error) {
-	if pallet, err = getPallet(c.String("workspace")); err != nil {
+	if pallet, err = getPallet(wpath); err != nil {
 		return nil, nil, err
 	}
-	if cache, _, err = fcli.GetRepoCache(c.String("workspace"), pallet, ensureCache); err != nil {
+	if cache, _, err = fcli.GetRepoCache(wpath, pallet, ensureCache); err != nil {
 		return nil, nil, err
 	}
 	return pallet, cache, nil
@@ -42,7 +42,7 @@ func getPallet(wpath string) (pallet *forklift.FSPallet, err error) {
 
 func cacheAllAction(versions Versions) cli.ActionFunc {
 	return func(c *cli.Context) error {
-		pallet, cache, err := processFullBaseArgs(c, false)
+		pallet, cache, err := processFullBaseArgs(c.String("workspace"), false)
 		if err != nil {
 			return err
 		}
@@ -63,7 +63,7 @@ func cacheAllAction(versions Versions) cli.ActionFunc {
 			fmt.Println("Done! No further actions are needed at this time.")
 			return nil
 		}
-		fmt.Println("Done! Next, you'll probably want to run `forklift plt stage`.")
+		fmt.Println("Done!")
 		return nil
 	}
 }
@@ -76,7 +76,9 @@ func switchAction(versions Versions) cli.ActionFunc {
 		if err != nil {
 			return err
 		}
-		pallet, repoCache, err := preparePallet(c, workspace, versions)
+		pallet, repoCache, err := preparePallet(
+			workspace, c.Args().First(), true, true, c.Bool("ignore-tool-version"), versions,
+		)
 		if err != nil {
 			return err
 		}
@@ -98,7 +100,8 @@ func switchAction(versions Versions) cli.ActionFunc {
 		}
 		if !c.Bool("apply") {
 			fmt.Println(
-				"Done! To apply the staged pallet immediately, run `sudo -E forklift stage apply`.",
+				"Done! To apply the staged pallet, you may need to reboot or run " +
+					"`forklift stage apply` (or `sudo -E forklift stage apply` if you need sudo for Docker).",
 			)
 			return nil
 		}
@@ -116,6 +119,9 @@ func switchAction(versions Versions) cli.ActionFunc {
 }
 
 func ensureWorkspace(wpath string) (*forklift.FSWorkspace, error) {
+	if !forklift.Exists(wpath) {
+		fmt.Printf("Making a new workspace at %s...", wpath)
+	}
 	if err := forklift.EnsureExists(wpath); err != nil {
 		return nil, errors.Wrapf(err, "couldn't make new workspace at %s", wpath)
 	}
@@ -130,70 +136,62 @@ func ensureWorkspace(wpath string) (*forklift.FSWorkspace, error) {
 }
 
 func preparePallet(
-	c *cli.Context, workspace *forklift.FSWorkspace, versions Versions,
+	workspace *forklift.FSWorkspace, gitRepoQuery string,
+	removeExistingLocalPallet, cacheStagingReqs, ignoreToolVersion bool, versions Versions,
 ) (pallet *forklift.FSPallet, repoCache forklift.PathedRepoCache, err error) {
 	// clone pallet
-	if err = os.RemoveAll(workspace.GetCurrentPalletPath()); err != nil {
-		return nil, nil, errors.Wrap(err, "couldn't remove local pallet")
+	if removeExistingLocalPallet {
+		fmt.Println("Warning: if a local pallet already exists, it will be deleted now...")
+		if err = os.RemoveAll(workspace.GetCurrentPalletPath()); err != nil {
+			return nil, nil, errors.Wrap(err, "couldn't remove local pallet")
+		}
 	}
 	if err = fcli.CloneQueriedGitRepoUsingLocalMirror(
-		0, workspace.GetPalletCachePath(), c.Args().First(), workspace.GetCurrentPalletPath(),
+		0, workspace.GetPalletCachePath(), gitRepoQuery, workspace.GetCurrentPalletPath(),
 	); err != nil {
 		return nil, nil, err
 	}
 	fmt.Println()
-	// TODO: warn if the git repo doesn't appear to be an actual pallet, or if the pallet's forklift
-	// version is incompatible
+	// TODO: warn if the git repo doesn't appear to be an actual pallet
 
-	// cache everything required by pallet
-	if pallet, repoCache, err = processFullBaseArgs(c, false); err != nil {
+	if pallet, repoCache, err = processFullBaseArgs(workspace.FS.Path(), false); err != nil {
 		return nil, nil, err
 	}
 	if err = fcli.CheckShallowCompatibility(
 		pallet, repoCache, versions.Tool, versions.MinSupportedRepo, versions.MinSupportedPallet,
-		c.Bool("ignore-tool-version"),
+		ignoreToolVersion,
 	); err != nil {
 		return pallet, repoCache, err
 	}
-	if _, err = fcli.CacheStagingRequirements(pallet, repoCache.Path()); err != nil {
-		return pallet, repoCache, err
+
+	// cache everything required by pallet
+	if cacheStagingReqs {
+		if _, err = fcli.CacheStagingRequirements(pallet, repoCache.Path()); err != nil {
+			return pallet, repoCache, err
+		}
 	}
 	return pallet, repoCache, nil
 }
 
 // clone
 
-func cloneAction(c *cli.Context) error {
-	wpath := c.String("workspace")
-	if !forklift.Exists(wpath) {
-		fmt.Printf("Making a new workspace at %s...", wpath)
-	}
-	if err := forklift.EnsureExists(wpath); err != nil {
-		return errors.Wrapf(err, "couldn't make new workspace at %s", wpath)
-	}
-	workspace, err := forklift.LoadWorkspace(wpath)
-	if err != nil {
-		return err
-	}
-	if err = forklift.EnsureExists(workspace.GetDataPath()); err != nil {
-		return errors.Wrapf(err, "couldn't ensure the existence of %s", workspace.GetDataPath())
-	}
-
-	if c.Bool("force") {
-		if err = os.RemoveAll(workspace.GetCurrentPalletPath()); err != nil {
-			return errors.Wrap(err, "couldn't remove local pallet")
+func cloneAction(versions Versions) cli.ActionFunc {
+	return func(c *cli.Context) error {
+		workspace, err := ensureWorkspace(c.String("workspace"))
+		if err != nil {
+			return err
 		}
-	}
-	if err := fcli.CloneQueriedGitRepoUsingLocalMirror(
-		0, workspace.GetPalletCachePath(), c.Args().First(), workspace.GetCurrentPalletPath(),
-	); err != nil {
-		return err
-	}
 
-	// TODO: warn if the git repo doesn't appear to be an actual pallet, or if the pallet's forklift
-	// version is incompatible
-	fmt.Println("Done! Next, you'll probably want to run `forklift plt cache-all`.")
-	return nil
+		if _, _, err = preparePallet(
+			workspace, c.Args().First(),
+			c.Bool("force"), !c.Bool("no-cache-req"), c.Bool("ignore-tool-version"), versions,
+		); err != nil {
+			return err
+		}
+
+		fmt.Println("Done!")
+		return nil
+	}
 }
 
 // fetch
@@ -220,26 +218,45 @@ func fetchAction(c *cli.Context) error {
 
 // pull
 
-func pullAction(c *cli.Context) error {
-	workspace, err := forklift.LoadWorkspace(c.String("workspace"))
-	if err != nil {
-		return err
-	}
-	palletPath := workspace.GetCurrentPalletPath()
+func pullAction(versions Versions) cli.ActionFunc {
+	return func(c *cli.Context) error {
+		workspace, err := forklift.LoadWorkspace(c.String("workspace"))
+		if err != nil {
+			return err
+		}
+		palletPath := workspace.GetCurrentPalletPath()
 
-	fmt.Println("Attempting to fast-forward the local pallet...")
-	updated, err := git.Pull(palletPath)
-	if err != nil {
-		return errors.Wrap(err, "couldn't fast-forward the local pallet")
-	}
-	if !updated {
-		fmt.Println("No changes from the remote release.")
-	}
+		fmt.Println("Attempting to fast-forward the local pallet...")
+		updated, err := git.Pull(palletPath)
+		if err != nil {
+			return errors.Wrap(err, "couldn't fast-forward the local pallet")
+		}
+		if !updated {
+			fmt.Println("No changes from the remote release.")
+		}
+		// TODO: display changes
 
-	// TODO: warn if the git repo doesn't appear to be an actual pallet, or if the pallet's forklift
-	// version is incompatible
-	// TODO: display changes
-	return nil
+		fmt.Println()
+
+		pallet, repoCache, err := processFullBaseArgs(c.String("workspace"), false)
+		if err != nil {
+			return err
+		}
+		if err = fcli.CheckShallowCompatibility(
+			pallet, repoCache, versions.Tool, versions.MinSupportedRepo, versions.MinSupportedPallet,
+			c.Bool("ignore-tool-version"),
+		); err != nil {
+			return err
+		}
+
+		if !c.Bool("no-cache-req") {
+			if _, err = fcli.CacheStagingRequirements(pallet, repoCache.Path()); err != nil {
+				return err
+			}
+		}
+		fmt.Println("Done!")
+		return nil
+	}
 }
 
 // rm
@@ -271,7 +288,7 @@ func showAction(c *cli.Context) error {
 
 func checkAction(versions Versions) cli.ActionFunc {
 	return func(c *cli.Context) error {
-		pallet, cache, err := processFullBaseArgs(c, true)
+		pallet, cache, err := processFullBaseArgs(c.String("workspace"), true)
 		if err != nil {
 			return err
 		}
@@ -293,7 +310,7 @@ func checkAction(versions Versions) cli.ActionFunc {
 
 func planAction(versions Versions) cli.ActionFunc {
 	return func(c *cli.Context) error {
-		pallet, cache, err := processFullBaseArgs(c, true)
+		pallet, cache, err := processFullBaseArgs(c.String("workspace"), true)
 		if err != nil {
 			return err
 		}
@@ -317,7 +334,7 @@ func planAction(versions Versions) cli.ActionFunc {
 
 func stageAction(versions Versions) cli.ActionFunc {
 	return func(c *cli.Context) error {
-		pallet, cache, err := processFullBaseArgs(c, true)
+		pallet, cache, err := processFullBaseArgs(c.String("workspace"), true)
 		if err != nil {
 			return err
 		}
@@ -354,7 +371,7 @@ func stageAction(versions Versions) cli.ActionFunc {
 
 func applyAction(versions Versions) cli.ActionFunc {
 	return func(c *cli.Context) error {
-		pallet, repoCache, err := processFullBaseArgs(c, true)
+		pallet, repoCache, err := processFullBaseArgs(c.String("workspace"), true)
 		if err != nil {
 			return err
 		}
